SymbolTable: Add fill_symbol_table and print_symbol_table functions
fill_symbol_table(std::ifstream&) makes a pass through the 6502 source
code and adds all the labels found to the symbol table with the label
name and the program counter when that label was encountered.

print_symbol_table() does precisely that. It is just a helpful tool for
debugging so that we can see all the values in the symbol table along
with the line at which it is created.