Tokenizer+SymbolTable: Add remove_comments() and remove_whitespace()
Both of these things need to happen to each line before it is tokenized
otherwise we tokenize unnecessary junk. remove_comments modifies the
line by reference and I am considering changed tokenize and remove_whi-
tespace() to do that as well. Makes for less variable delcarations but
could lead to confusion as we have multiple functions modifiying one
line. Now that I think about it, it is probably best NOT to pass by
reference as it is pretty dangerous to modify our source code. Think
about if something was broken in remove_whitespace() and instead of
it just screwing up a copy of the line, it screws up the real thing.
Thinking out loud.