OpCodes: Switch from enum to map for instruction set
THIS IS NOT SET IN STONE AND MAY CHANGE. I believe at the moment the
best option is a map with a standardized instruction form. Lines in the
source code will be translated to the standardized version and then
looked up in the map to get the corresponding hex code which will then
be translated to binary and inserted to the output file.