Translate+SymbolTable: Add standardize_instruction function
Standardize_function() is a way to put the current instruction into an
easily mappable form. Only the information needed in order to select
the right instruction is included and specified. The other information
is either turned to a variable, X, or removed entirely, like when we
encounter a '#'.

With the standardized_function we can map it to a hex opcode. The
address will be pulled out and saved before standardizing so that we
can convert it to hex and insert it into the binary after inserting the
instruction op code.