Translate: Add translate_instruction_to_hex_opcode() function
Does precisely as the name says. It looks through our OpCode map and
returns the value that corresponds to the instruction that was passed
as a parameter. If a key that matches the instruction is not found, it
returns and arbitrary 45.

This is the first step towards translating instructions on a larger
scale :). Going to crack away at this this weekend to hopefully get
most of the 6502 instruction set added to the OpCode map and then be
able to convert instructions down to the correct key.