TestTranslate: Get tests working and implement first part of test
So it begins. This is the initial commit for the Translate/ tests. The
CMakeLists.txt file works and we can run it and then make to compile
the tests into an executable. All this should be done in a script which
will get implemented soon. It should run test cases, then build the
system.