Tokenizer+Tests: Add index_of_oper function to tokenizer + test it
index_of_oper(std::vector<std::string>) returns the index of the oper
in the tokenized instruction. This is important for being able to
separate this value away from the instruction before the instruction is
standardized.

Also wrote three test cases for index_of_oper. Two were normal and one
is an edge case in which it does not find an oper in the instruction.
If an oper is not found it returns -1 which is not a valid index and
can easily be checked against to prevent bounds errors.