Translate: Change substring to its correct usage
It has been brought to my attention (thank you Competitive Prgrammer's
Handbook) that I did not know how substring worked... Substring takes
the starting index and then the length of the desired substring. Not
the starting index and the ending index.

This allows for me to correct the oper_high_byte() function and
explains why the oper_high_byte() function was not giving me the substr
I wanted it to. Because I was using substr wrong.

All Translate tests passed. :)