Tokenizer+Tests: Make oper its own function and add function for index
Before, we were actually getting the index of the character in the
string in the tokenized line. We don't want that. Instead we want to
know which token holds the address, and then from that token derive the
address itself.

This commit fixes that. index_of_oper_in_tokenized line returns which
token holds the index. Oper() then looks in that token and extracts the
address (no #$ included, just the numbers). Then oper_low_byte() and
oper_high_byte() can do their thing without having to deal with any
kind of trimming or searching. Cleaner all around, though I do realize
its still not very clean.