Tokenizer: Handle indirect addressing start index for oper substring
Test was failing because our index was being set incorrectly if we had
something like "($9382)". It was setting it one too far because it
the start index was set to 3 if it saw a parentheses but this is not
always the case. If there is a parentheses and then a # then it is 3
but if a $ follows the parentheses, it should only be 2.