Translate+Tests: Change address_as_int to return hexadecimal
Needs to return hexadecimal instead of decimal otherwise we convert a
decimal value to binary instead of a hexadecimal number. It is in decimal,
but it returns the hex value (ie if pass 0x98 it returns 152).

This is done by use stoi(string, 0, 16) when returning the value.

Tests have been modified to match this behavior and all are passing.
I think this should probably be rewritten because I'm confused just
writing this commit. Renaming would be REALLY helpful.