SymbolTable: Rewrite, read description
Make functions static as we only deal with static variables in the
SymbolTable class. Change the function calls in Assemble to use static
method calls. Mark print_symbol_table as maybe_unused due it being more
of a debugging function that is in fact maybe unused. Change
constructor and destructor to default and don't call them in cpp file.
Rename line to buffer because it is a buffer. Make most parameters
const/const&.