Assemble: Add memory_address variable
Both memory address and program counter need to be tracked but their
values should be different. Therefore we can't use just the program
counter for both. This adds a memory_address variable that keeps track
of memory starting from 0x800. Use this value for the hexdumps not the
program counter.