Assemble+Translate: Correctly ignore blank lines
Add is_blank_line(string) function to check if the given buffer just
consists of characters or a newline. Then call the function in Assemble
along with checking if the buffer is empty. In any of the three cases,
is_label, is_blank_line, or buffer.empty(), continue to the next line
of the file.