Everywhere: Use optionals instead of random return values
Instead of returning -1 when a value is not present, return an optional
so that it can just be checked as has_value(). Optional is what I
wanted all along just did not know it existed/how to use it. Much nicer
to return an optional and check it's value than to return some
arbitrary number and hope that number doesn't actually come along.