Everywhere: Remove TranslationHelpers class
Move all functionality of TranslationHelpers class to Translate and
make as many functions private as possible. No reason to have a whole
other class just for some helper functions. Ideally, they would all be
private but they are used in Assembly.cpp which messes that up.

Rip out test for the function that is made private. All tests pass.